+++ src/middleware/InjectableResolver.ts
@@ -1,0 +1,5 @@
+import { MiddlewareContext, MiddlewareResolver } from './resolver';
+
+export class InjectableResolver implements MiddlewareResolver {
+    resolveNext<T>(context: MiddlewareContext<T>): T {}
+}